# frontend/src/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Neuro Eye AI",
    "description": "\nNeuroEyeAI REST API for eye-tracking analysis and multi-tenant user management.\n\n## Authentication\n\nMost endpoints require a JWT bearer token. Use the **Authorize** button in the\ndocumentation UI to set your token, or send:\n\n```\nAuthorization: Bearer <your-token>\n```\n",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/health": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Health check",
        "description": "Returns service status. Use for load balancers and uptime monitoring.",
        "operationId": "health_check",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Check"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register",
        "operationId": "register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Current user",
        "operationId": "get_me",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthUserResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout",
        "operationId": "logout",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Forgot password",
        "operationId": "forgot_password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ForgotPasswordRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Reset password",
        "operationId": "reset_password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResetPasswordRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients": {
      "get": {
        "tags": ["Patients"],
        "summary": "List patients",
        "operationId": "get_patients",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "query", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "page_size", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PatientListResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Patients"],
        "summary": "Create patient",
        "operationId": "create_patient",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PatientCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PatientResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/{patient_id}": {
      "get": {
        "tags": ["Patients"],
        "summary": "Get patient",
        "operationId": "get_patient_by_id",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "patient_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PatientResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboard/summary": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Dashboard KPIs",
        "operationId": "get_dashboard_summary",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DashboardSummaryResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/presets": {
      "get": {
        "tags": ["Vision"],
        "summary": "List test presets",
        "operationId": "get_test_presets",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/TestPresetResponse" }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions": {
      "get": {
        "tags": ["Sessions"],
        "summary": "List sessions",
        "operationId": "get_sessions",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "patient_id", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "page_size", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionListResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Sessions"],
        "summary": "Create session",
        "operationId": "create_session",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SessionCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}": {
      "get": {
        "tags": ["Sessions"],
        "summary": "Get session",
        "operationId": "get_session_by_id",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "session_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/report": {
      "get": {
        "tags": ["Reports"],
        "summary": "Get diagnostic report",
        "operationId": "get_session_report",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "session_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionReportResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthUserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "tenant": {
            "type": "string",
            "title": "Tenant"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "role",
          "tenant",
          "tenant_id"
        ],
        "title": "AuthUserResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserResponse"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user"
        ],
        "title": "LoginResponse"
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "password", "first_name", "last_name"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 8 },
          "first_name": { "type": "string" },
          "last_name": { "type": "string" },
          "second_last_name": { "type": "string", "nullable": true }
        }
      },
      "ForgotPasswordRequest": {
        "type": "object",
        "required": ["email"],
        "properties": { "email": { "type": "string", "format": "email" } }
      },
      "ResetPasswordRequest": {
        "type": "object",
        "required": ["token", "password"],
        "properties": {
          "token": { "type": "string" },
          "password": { "type": "string", "minLength": 8 }
        }
      },
      "MessageResponse": {
        "type": "object",
        "required": ["message"],
        "properties": { "message": { "type": "string" } }
      },
      "PatientCreate": {
        "type": "object",
        "required": ["first_name", "last_name"],
        "properties": {
          "first_name": { "type": "string" },
          "last_name": { "type": "string" },
          "second_last_name": { "type": "string", "nullable": true },
          "email": { "type": "string", "nullable": true },
          "phone_number": { "type": "string", "nullable": true },
          "sex": { "type": "string", "nullable": true },
          "notes": { "type": "string", "nullable": true }
        }
      },
      "PatientResponse": {
        "type": "object",
        "required": ["id", "first_name", "last_name", "display_name", "risk_level"],
        "properties": {
          "id": { "type": "string" },
          "first_name": { "type": "string" },
          "last_name": { "type": "string" },
          "second_last_name": { "type": "string", "nullable": true },
          "display_name": { "type": "string" },
          "birth_date": { "type": "string", "nullable": true },
          "email": { "type": "string", "nullable": true },
          "phone_number": { "type": "string", "nullable": true },
          "sex": { "type": "string", "nullable": true },
          "notes": { "type": "string", "nullable": true },
          "risk_level": { "type": "string" },
          "session_count": { "type": "integer" }
        }
      },
      "PatientListResponse": {
        "type": "object",
        "required": ["items", "total", "page", "page_size"],
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/PatientResponse" } },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "page_size": { "type": "integer" }
        }
      },
      "DashboardSummaryResponse": {
        "type": "object",
        "required": ["patient_count", "session_count", "completed_session_count", "at_risk_count"],
        "properties": {
          "patient_count": { "type": "integer" },
          "session_count": { "type": "integer" },
          "completed_session_count": { "type": "integer" },
          "at_risk_count": { "type": "integer" }
        }
      },
      "TestPresetResponse": {
        "type": "object",
        "required": ["id", "code", "name", "config"],
        "properties": {
          "id": { "type": "string" },
          "code": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "config": { "type": "object" }
        }
      },
      "SessionCreate": {
        "type": "object",
        "required": ["patient_id", "preset_id"],
        "properties": {
          "patient_id": { "type": "string" },
          "preset_id": { "type": "string" }
        }
      },
      "SessionMetricsResponse": {
        "type": "object",
        "required": ["session_id", "sample_count", "mean_fps", "detection_rate", "mean_latency_ms", "fixation_stability", "saccade_amplitude", "extras", "risk_level"],
        "properties": {
          "session_id": { "type": "string" },
          "sample_count": { "type": "integer" },
          "mean_fps": { "type": "number" },
          "detection_rate": { "type": "number" },
          "mean_latency_ms": { "type": "number" },
          "fixation_stability": { "type": "number" },
          "saccade_amplitude": { "type": "number" },
          "extras": { "type": "object" },
          "risk_level": { "type": "string" }
        }
      },
      "SessionResponse": {
        "type": "object",
        "required": ["id", "patient_id", "patient_name", "clinician_id", "preset_id", "preset_code", "preset_name", "status"],
        "properties": {
          "id": { "type": "string" },
          "patient_id": { "type": "string" },
          "patient_name": { "type": "string" },
          "clinician_id": { "type": "string" },
          "preset_id": { "type": "string" },
          "preset_code": { "type": "string" },
          "preset_name": { "type": "string" },
          "status": { "type": "string" },
          "started_at": { "type": "string", "nullable": true },
          "ended_at": { "type": "string", "nullable": true },
          "abort_reason": { "type": "string", "nullable": true },
          "metrics": { "$ref": "#/components/schemas/SessionMetricsResponse" }
        }
      },
      "SessionListResponse": {
        "type": "object",
        "required": ["items", "total", "page", "page_size"],
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/SessionResponse" } },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "page_size": { "type": "integer" }
        }
      },
      "SessionReportResponse": {
        "type": "object",
        "required": ["session", "disclaimer"],
        "properties": {
          "session": { "$ref": "#/components/schemas/SessionResponse" },
          "metrics": { "$ref": "#/components/schemas/SessionMetricsResponse" },
          "disclaimer": { "type": "string" }
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter a JWT access token."
      }
    }
  },
  "tags": [
    {
      "name": "Health",
      "description": "Service health and readiness checks."
    },
    {
      "name": "Auth",
      "description": "Authentication and session management."
    },
    {
      "name": "Patients",
      "description": "Patient records for the authenticated tenant."
    },
    {
      "name": "Sessions",
      "description": "Screening sessions, gaze metrics, and history."
    },
    {
      "name": "Dashboard",
      "description": "Clinical overview KPIs."
    },
    {
      "name": "Reports",
      "description": "Diagnostic reports generated from completed sessions."
    },
    {
      "name": "Vision",
      "description": "Eye-tracking and vision analysis endpoints."
    }
  ]
}